GET WIFI PASSWORD
Introduction :
Everyone requires a stopwatch, to calculate the time spent in any participation or some other cases. Every electronic device is included with a pre-installed stopwatch, but what if I tell you that, you can achieve the countdown timer just by a few lines of python code on your own, interesting right? In this tutorial, we’ll look into how we can achieve a countdown timer in python.

Outline :
A countdown timer is just a simple clock that counts the time in reverse from the starting point to zero.

A countdown timer is a virtual clock on the console that counts down from a certain number to indicate the beginning or end of an event or offer. This rushes the people to do the actions a quick and generates more conversions because “time is running out.” Countdown timers can also be used to count down to when an offer becomes available at a specific date and time.

Project Prerequisites : 
Time: The Python time module provides many ways of representing time in code, such as objects, numbers, and strings. It also provides functionality other than representing time, like waiting during code execution and measuring the efficiency of your code.

That’s it! You don’t have to install any other packages for this project.