Create COUNTDOWN TIMER
Outline :
A countdown timer is just a simple clock that counts the time in reverse from the starting point to zero.

A countdown timer is a virtual clock on the console that counts down from a certain number to indicate the beginning or end of an event or offer. This rushes the people to do the actions a quick and generates more conversions because “time is running out.” Countdown timers can also be used to count down to when an offer becomes available at a specific date and time.

Project Prerequisites : 
Time: The Python time module provides many ways of representing time in code, such as objects, numbers, and strings. It also provides functionality other than representing time, like waiting during code execution and measuring the efficiency of your code.

That’s it! You don’t have to install any other packages for this project.

Code Implementation : 
To understand this project, you should know about the following Python programming topics:

● Python while Loop
● Python divmod()
● Python time Module

The first thing we need to do is to import the time module.

#import time module for the further usage 
import time
We defined a function named countdown () and passed the input time as an argument.

The divmod() method takes two numbers and returns a pair of numbers in the tuple data structure consisting of their quotient and remainder.

end=’\r‘ overwrites the output for each iteration, which gives the perfect look of a countdown timer.

The value of t, which is represented as an input of time, is decremented at the end of each iteration.

Python time method sleep() suspends execution for the given number of seconds. The argument may be a floating-point number to indicate a more accurate delay of time in each iteration.

The line in the code, {:02d}:{:02d}’.format(mins, secs) Indicates that there should be two in mins and two digits in seconds as the mins and secs are represented in the format() function.

A while() loop is used to iterate the code until it becomes zero, which is the final destination of the countdown timer.



here while loop is implemented as it continues until the time is greater than or equal to zero.

#while loop implementation
    while t:
        mins, secs = divmod(t, 60)
        timer = '{:02d}:{:02d}'.format(mins, secs)
        print(timer, end="\r")
        time.sleep(1)
        t -= 1